# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payment API (India)",
    "description": "Payment processing API for India – UPI, cards, net banking, wallets, cash on delivery. TrendRush-style schema (raw MySQL).",
    "version": "1.0.0",
    "contact": { "name": "API Support" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
...
  "paths": {
...
    "/api/payments/create": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create payment",
        "description": "Creates payment (TrendRush schema). Requires order to exist in norders and billing/shipping addressId.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "currency", "paymentMethod", "orderId", "userId", "customerPhone", "billingAddress", "shippingAddress"],
                "properties": {
                  "amount": { "type": "number", "example": 100 },
                  "currency": { "type": "string", "example": "INR" },
                  "paymentMethod": { "type": "string", "enum": ["upi", "card", "netbanking", "wallet", "cash_on_delivery"] },
                  "orderId": { "type": "string", "description": "Order number (e.g. ORD_...) or numeric order_id" },
                  "userId": { "type": "string" },
                  "customerPhone": { "type": "string", "example": "+919876543210" },
                  "customerEmail": { "type": "string" },
                  "customerName": { "type": "string" },
                  "billingAddress": {
                    "type": "object",
                    "required": ["addressId"],
                    "properties": {
                      "addressId": { "type": "integer" },
                      "state": { "type": "string" }
                    }
                  },
                  "shippingAddress": {
                    "type": "object",
                    "required": ["addressId"],
                    "properties": {
                      "addressId": { "type": "integer" },
                      "state": { "type": "string" }
                    }
                  },
                  "items": { "type": "array", "items": { "type": "object", "properties": { "name": {}, "quantity": {}, "price": {} } } },
                  "gstNumber": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Payment created" },
          "400": { "description": "Validation or business error" }
        }
      }
    },
    "/api/payments/confirm": {
      "post": {
        "tags": ["Payments"],
        "summary": "Confirm payment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["paymentId", "paymentMethod"],
                "properties": {
                  "paymentId": { "type": "integer" },
                  "paymentMethod": { "type": "string", "enum": ["upi", "card", "netbanking", "wallet", "cash_on_delivery"] },
                  "paymentData": {
                    "type": "object",
                    "properties": {
                      "razorpayOrderId": { "type": "string" },
                      "razorpayPaymentId": { "type": "string" },
                      "razorpaySignature": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment confirmed" },
          "400": { "description": "Verification failed" },
          "404": { "description": "Payment not found" }
        }
      }
    },
    "/api/payments/status/{paymentId}": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get payment status",
        "parameters": [
          { "name": "paymentId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Payment details and status" },
          "404": { "description": "Payment not found" }
        }
      }
    },
...
    "/api/orders/create": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "items", "totalAmount", "shippingAddress", "billingAddress"],
                "properties": {
                  "userId": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": { "type": "object", "properties": { "name": {}, "quantity": {}, "price": {} } }
                  },
                  "totalAmount": { "type": "number" },
                  "shippingAddress": { "type": "object" },
                  "billingAddress": { "type": "object" },
                  "paymentMethod": { "type": "string" },
                  "gstNumber": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created; returns orderId (order_number), order_id, shippingAddressId, billingAddressId"
          },
          "400": { "description": "Validation failed" }
        }
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID or order number",
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Order number (ORD_...) or numeric order_id" }
        ],
        "responses": {
          "200": { "description": "Order and payment (if any)" },
          "404": { "description": "Order not found" }
        }
      }
    },
...
  }
}